Core.SymbolTree

Types

data Symbol nt t

The sum type (disjoint union) of non-terminal and terminal.

nt the non-terminal type

t the terminal type

Constructors

T t

NT nt

data SymbolTree nt t

SymbolTree nt t encodes a traditional parse tree for a context free grammar.

nt node label type (non-terminal)

t leaf label type (temrinal)

It is a more restricted type than Tree (Symbol nt t) in that the leaf node are always of type t while branching node content always of type nt.

Constructors

TLeaf t

NTNode nt [SymbolTree nt t]

Query

rootNT :: SymbolTree nt t -> Maybe nt

root non-terminal of a SymbolTree. If the symbolTree is just a terminal leaf, it returns Nothing.

rootSymbol :: SymbolTree nt t -> Symbol nt t

root symbol of a SymbolTree

Conversion

symbolTreetoTree :: SymbolTree nt t -> Tree (Symbol nt t)

(For fold/unfold) Base functor from recursion scheme

data SymbolTreeF nt t r

Base functor for SymbolTree nt t

Fix (SymbolTreeF nt t) ≃ SymbolTree nt t

Constructors

TLeafF t

NTNodeF nt [r]